🏗️ Move ProfileAdapter, ProfileFragment, ProfileViewModel to presenter layer
